Project Commitment
In this commit I add the software part for the project. Firstly it contains 4 files which perform the minimun requirements asked for the project: adc, pwm, serial communication and gpio (buttons). Secondly there are another three files which implements the logic for the disable Alarm system. One of the implements the system performance wihtout serial communication with other device, the second one implements also the serial communication with de device with the PIR sensor and finally the third one implements low power mode. Deeper description of the files are explained  in the README file.
Author: Eduardo (jedur18)